feat: modularize database operations
